finished checking error messages, and routes for review
